SSCS-10192 created utility class
