+++ src/main/java/uk/gov/hmcts/reform/sscs/helper/mapping/HearingsDetailsMapping.java
@@ -1,8 +1,8 @@
 package uk.gov.hmcts.reform.sscs.helper.mapping;
 
 
-import org.jetbrains.annotations.NotNull;
 import uk.gov.hmcts.reform.sscs.ccd.domain.*;
+import uk.gov.hmcts.reform.sscs.helper.mappingutils.GetVenueMultipleEpims;
 import uk.gov.hmcts.reform.sscs.model.HearingWrapper;
 import uk.gov.hmcts.reform.sscs.model.single.hearing.*;
 import uk.gov.hmcts.reform.sscs.model.single.hearing.HearingDetails;
@@ -10,17 +10,14 @@ import uk.gov.hmcts.reform.sscs.model.single.hearing.HearingDetails;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
-import java.util.Objects;
 import java.util.stream.Collectors;
 
 import static java.util.Objects.nonNull;
 import static org.apache.commons.lang3.StringUtils.isNotBlank;
 import static uk.gov.hmcts.reform.sscs.ccd.domain.YesNo.isYes;
 
-@SuppressWarnings({"PMD.UnnecessaryLocalBeforeReturn","PMD.ReturnEmptyCollectionRatherThanNull"})
+@SuppressWarnings({"PMD.UnnecessaryLocalBeforeReturn","PMD.ReturnEmptyCollectionRatherThanNull", })
 // TODO Unsuppress in future
 public final class HearingsDetailsMapping {
 
@@ -29,9 +26,6 @@ public final class HearingsDetailsMapping {
     public static final int DURATION_SESSIONS_MULTIPLIER = 165;
     public static final int DURATION_HOURS_MULTIPLIER = 60;
     public static final int DURATION_DEFAULT = 30; // TODO find out default
-    public static final String Manchester = "Manchester";
-    public static final String Chester = "Chester";
-    public static final String Plymouth = "Plymouth";
 
     private HearingsDetailsMapping() {
 
@@ -161,64 +155,16 @@ public final class HearingsDetailsMapping {
         return false;
     }
 
-    private static List<HearingLocations> getMultipleLocationDetails(CaseManagementLocation caseManagementLocation) {
-
-        Map<String, List<String>> epimMap = new HashMap<>();
-        List<String> chesterId = new ArrayList<>(List.of("226511", "443014"));
-        List<String> manchesterId = new ArrayList<>(List.of("512401", "701411"));
-        List<String> plymouthId = new ArrayList<>(List.of("764728", "235590"));
-
-        epimMap.put(Manchester, manchesterId);
-        epimMap.put(Chester, chesterId);
-        epimMap.put(Plymouth, plymouthId);
-
-        List<HearingLocations> locationId = new ArrayList<>();
-        String processingCenter = caseManagementLocation.getRegion();
-        HearingLocations hearingLocations = new HearingLocations();
-        hearingLocations.setLocationId(caseManagementLocation.getBaseLocation());
-        hearingLocations.setLocationType(processingCenter);
-        switch (processingCenter) {
-            case "Manchester": locationId.addAll(getEpims(epimMap, Manchester, hearingLocations));
-                break;
-            case "Chester": locationId.addAll(getEpims(epimMap, Chester, hearingLocations));
-                break;
-            case "Plymouth": locationId.addAll(getEpims(epimMap, Plymouth, hearingLocations));
-                break;
-            default: break;
-        }
-        return locationId;
-    }
-
-    @NotNull
-    private static List<HearingLocations> getEpims(Map<String, List<String>> epimLists, String locationName, HearingLocations hearingLocations) {
-        List<List<String>> epims;
-        List<HearingLocations> epimss = new ArrayList<>();
-
-        epims = epimLists
-            .entrySet()
-            .stream()
-            .filter(e -> Objects.equals((e.getKey()), locationName))
-            .map(Map.Entry::getValue)
-            .collect(Collectors.toList());
-
-        epims.forEach(epimId -> {
-            if (epimId.contains(hearingLocations.getLocationId())) {
-                hearingLocations.setMultipleLocationId(epimId);
-                epimss.add(hearingLocations);
-            }
-        });
-        return epimss;
-    }
-
     public static List<HearingLocations> getHearingLocations(CaseManagementLocation caseManagementLocation) {
         HearingLocations hearingLocations = new HearingLocations();
+        GetVenueMultipleEpims venueMultipleEpims = new GetVenueMultipleEpims();
         List<HearingLocations> location;
-        List<String> multipleLocationList = new ArrayList<>(List.of(Manchester, Chester, Plymouth));
+        List<String> multipleLocationList = new ArrayList<>(List.of("Manchester", "Chester", "Plymouth"));
 
         List<HearingLocations> hearingLocationsList = new ArrayList<>();
 
-        if (multipleLocationList.contains(caseManagementLocation.getBaseLocation())) {
-            location = getMultipleLocationDetails(caseManagementLocation);
+        if (multipleLocationList.contains(caseManagementLocation.getRegion())) {
+            location = venueMultipleEpims.getMultipleLocationDetails(caseManagementLocation);
             hearingLocationsList.addAll(location);
         } else {
             hearingLocations.setLocationId(caseManagementLocation.getBaseLocation());
